feat(react): Add basic version of session-plan and learning goals for Week 3
I've changed the topic from "Lifecycle" to "Hooks", because the learning
materials did not provide a lot for "Lifecycle", but talk about Hooks
extensively. They are part of the lifecycle concept, but I currently
think that it's more important for the trainees to understand hooks
first and know how to use them. Then the lifecycle will become clear.